UIE-204 App Startup - unit tests pt1 - initial code shift to callable (and testable) functions.
